service_api/flask/requirements.txt
click==7.1.2
Flask==1.1.2
Flask-MySQL==1.5.1
itsdangerous==1.1.0
Jinja2==2.11.2
json-lxml==0.1.4
MarkupSafe==1.1.1
PyMySQL==0.10.0
python-dotenv==0.14.0
requests==2.24.0
Werkzeug==1.0.1
cryptography==2.9.2